\name{StateAR3}
\alias{StateAR3}

\title{
  A List of the MCMC Output Based on The GLMM-AR(p) Model
}
\description{
  This list contains the mcmc output from the R function of GLMMARp.Binary().
}
\usage{
data(StateAR2)
}
\format{
 The mcmc output as a list contains the posterior distributions of all the parameters 
 in the GLMM-AR(p) model, produced by the function of GLMMARp.Binary().
}

\source{
 Original data provided by the author of the GLMMarp Package.
}

\keyword{datasets}
