\name{StateFailure}
\alias{StateFailure}
\title{
  Sub-Suharan State Failure Study Data 
}
\description{
  This dataframe contains a matrix in which is a TSCS data set about state failure events
  in the sub-Saharan countries
}
\usage{
data(StateFailure)
}
\format{
  The data set is based on the Political Instability Task Force (PITF) 2001 data base. The data 
  used here only contain the observations in the sub-Saharan countries. There are many missing data 
  in the original data base, but missingness in this used data set has already been handled with 
  the multiple imputation method. 
}

\source{
  Political Instability Task Force. \emph{PITF Phase III Findings.} \url{http://globalpolicy.gmu.edu/pitf/}.
}

\keyword{datasets}
